Riccardo4

+++ HackHub/src/main/java/it/unicam/cs/ids/hackhub/dto/hackathon/HackathonRegistrationsDTO.java
@@ -3,4 +3,7 @@ package it.unicam.cs.ids.hackhub.dto.hackathon;
 import it.unicam.cs.ids.hackhub.dto.team.TeamSummaryDTO;
 import java.util.List;
 
-public record HackathonRegistrationsDTO(long totalRegistrations, List<TeamSummaryDTO> teams) {}
+
+public record HackathonRegistrationsDTO(int totalRegistrations,
+                                        List<TeamSummaryDTO> teams) {
+}
